Appointment

+++ SEP490_HealthEA/Infrastructure/MediatR/Events/Commands/DeleteEvent/DeleteAllEvent.cs
@@ -8,6 +8,7 @@ namespace Infrastructure.MediatR.Events.Commands.DeleteEvent;
 public class DeleteAllEventCommand : IRequest<bool>
 {
     public Guid OriginalEventId { get; set; } = Guid.NewGuid();
+    public Guid UserId { get; set; }
 }
 
 
@@ -23,33 +24,42 @@ public class DeleteAllEventCommandHandler : IRequestHandler<DeleteAllEventComman
 
     public async Task<bool> Handle(DeleteAllEventCommand request, CancellationToken cancellationToken)
     {
-        try
+        cancellationToken.ThrowIfCancellationRequested();
+
+        var eventsToDelete = await _context.Events
+            .Where(e => e.OriginalEventId == request.OriginalEventId)
+            .Include(e => e.Reminders)
+            .ToListAsync(cancellationToken);
+
+        if (!eventsToDelete.Any())
         {
-            var eventsToDelete = await _context.Events
-                .Where(e => e.OriginalEventId == request.OriginalEventId)
-                .Include(e => e.Reminders)
-                .ToListAsync(cancellationToken);
-
-            if (!eventsToDelete.Any())
-            {
-                throw new Exception(ErrorCode.EVENT_NOT_FOUND);
-            }
-
-            foreach (var eventToDelete in eventsToDelete)
-            {
-                _context.Reminders.RemoveRange(eventToDelete.Reminders);
-            }
-
-            _context.Events.RemoveRange(eventsToDelete);
-
-            await _context.SaveChangesAsync(cancellationToken);
-
-            return true; 
+            throw new Exception(ErrorCode.EVENT_NOT_FOUND);
         }
-        catch (Exception ex)
+
+        var eventIdsToDelete = eventsToDelete.Select(e => e.EventId).ToList();
+
+        var userEvents = await _context.UserEvents
+            .Where(ue => eventIdsToDelete.Contains(ue.EventId) && ue.UserId == request.UserId)
+            .ToListAsync(cancellationToken);
+
+        if (userEvents.Count == 0)
         {
-            return false;
+            throw new Exception(ErrorCode.UNAUTHORIZED_ACCESS);
         }
+
+        _context.UserEvents.RemoveRange(userEvents);
+
+        foreach (var eventToDelete in eventsToDelete)
+        {
+            _context.Reminders.RemoveRange(eventToDelete.Reminders);
+        }
+
+        _context.Events.RemoveRange(eventsToDelete);
+
+        await _context.SaveChangesAsync(cancellationToken);
+
+        return true;
     }
+
 
 }
