+++ SEP490_HealthEA/Infrastructure/MediatR/Events/Commands/DeleteEvent/DeleteEvent.cs
@@ -8,6 +8,7 @@ namespace Infrastructure.MediatR.Events.Commands.DeleteEvent;
 public class DeleteEventCommand : IRequest<bool>
 {
     public Guid EventId { get; set; }
+    public Guid UserId { get; set; }
 }
 
 public class DeleteEventCommandHandler : IRequestHandler<DeleteEventCommand, bool>
@@ -21,18 +22,25 @@ public class DeleteEventCommandHandler : IRequestHandler<DeleteEventCommand, boo
 
     public async Task<bool> Handle(DeleteEventCommand request, CancellationToken cancellationToken)
     {
-        cancellationToken.ThrowIfCancellationRequested();
-
         var eventEntity = await _context.Events
-            .Include(e => e.Reminders) // Include related reminders to delete them as well
-            .FirstOrDefaultAsync(e => e.EventId == request.EventId, cancellationToken);
+                .Include(e => e.Reminders)
+                .FirstOrDefaultAsync(e => e.EventId == request.EventId, cancellationToken);
 
         if (eventEntity == null)
             throw new Exception(ErrorCode.EVENT_NOT_FOUND);
 
-        // Remove the event and its related reminders
-        _context.Events.Remove(eventEntity);
+        var userEvents = await _context.UserEvents
+            .Where(ue => ue.EventId == request.EventId && ue.UserId == request.UserId)
+            .ToListAsync(cancellationToken);
 
+        if (userEvents.Count ==0)
+        {
+            throw new Exception(ErrorCode.UNAUTHORIZED_ACCESS);
+        }
+        _context.UserEvents.RemoveRange(userEvents);
+        await _context.SaveChangesAsync(cancellationToken);
+
+        _context.Events.Remove(eventEntity);
         await _context.SaveChangesAsync(cancellationToken);
 
         return true;
